change the return type to Tuple
This is standard practice; functions input and output tuples. There is no reason to make the returned Tuple into a Vector, and if anything it is bad practice; it makes a type unstable vector and it serves no purpose. Anything you could do with this two element vector you could do better and faster with a Tuple.