first pass at performance: complete refactor of functions
The point is: local estimation should be its own function because it is a dedicated piece of algorithm. Then, we can optmize stuff by looping over each "column" of the distance matrix. If it fits in memory, we also gain massive BLAS-enabled performance. If not, we can still optimize by pre-allocating vectors and not re-allocated memory in each pass.