add extremesdim (#9)
* Add files via upload

* Delete FractalDimensions.jl

* add extremesdim

* changed misspelled persistence

* remove annoying folder
 Please enter the commit message for your changes. Lines starting

* turn the initial string into comment

* change the return type to Tuple

This is standard practice; functions input and output tuples. There is no reason to make the returned Tuple into a Vector, and if anything it is bad practice; it makes a type unstable vector and it serves no purpose. Anything you could do with this two element vector you could do better and faster with a Tuple.

* Delete all `;`

No instance of `;` in the source code had reason to exist. Julia is capable of automatically deducing when a statement ends. Only reason for `;` is to have to commands on one line like:
`x = 3; y = x + 2`.

* organize import and export statements

* Correct structure in docstrings

* replace quanti with q

Also, make the error messages much shorter

* remove space from all `:: ` 

Apparently it is possible to have space between `::` but it is very un-idiomatic; noone does it in real code, which makes putting spaces hard to read.

* replace `AbstractVector` with `AbstractStateSpaceSet`

This is arguably the most important change I do in this PR. Read the docstring of `StateSpaceSet` to understand why.

* Fix the `extremal_index_sueveges` function...

There were a lot of problems:

- the function name was different from where it was called
- the function input arguments were named differently from where the function was called. This was confusing.
- The order of arguments and the documentation string where wrong. This was ULTRA confusing. The threshold and quantile arguments were exchanging places in the documentation.
- The quantile and threshold were named the same way. Not possible to deduce which was what besides going through the algorithm.

* rename variables for clarity

* don't call extremal index; function is wrong

* use a normal way to get length of set

* add convenience function for dim

* first pass at performance: complete refactor of functions

The point is: local estimation should be its own function because it is a dedicated piece of algorithm. Then, we can optmize stuff by looping over each "column" of the distance matrix. If it fits in memory, we also gain massive BLAS-enabled performance. If not, we can still optimize by pre-allocating vectors and not re-allocated memory in each pass.

* add convenience function for local values

* add progress bar

and also stop using `AbstractStateSpaceSet` in function call signatures

* add test file

* correctly use greater-equals instead of only greater

* add second estimator for pareto sigma

* parallelize via threads

* add preliminary tests for extreme value dim

* fix test for extremal index

* wip] add type for deciding estimator (?)

* we don't need a type in the end.

* rename q to p as it quantifies a probability, not the quantile

* wip evt

* finish docs EVT

* add passing tests to EVT

* add to docs

* changelog

---------

Co-authored-by: Datseris <datseris.george@gmail.com>