Block Maxima (#30)
* Add files via upload

* Delete FractalDimensions.jl

* add extremesdim

* changed misspelled persistence

* remove annoying folder
 Please enter the commit message for your changes. Lines starting

* turn the initial string into comment

* change the return type to Tuple

This is standard practice; functions input and output tuples. There is no reason to make the returned Tuple into a Vector, and if anything it is bad practice; it makes a type unstable vector and it serves no purpose. Anything you could do with this two element vector you could do better and faster with a Tuple.

* Delete all `;`

No instance of `;` in the source code had reason to exist. Julia is capable of automatically deducing when a statement ends. Only reason for `;` is to have to commands on one line like:
`x = 3; y = x + 2`.

* organize import and export statements

* Correct structure in docstrings

* replace quanti with q

Also, make the error messages much shorter

* remove space from all `:: ` 

Apparently it is possible to have space between `::` but it is very un-idiomatic; noone does it in real code, which makes putting spaces hard to read.

* replace `AbstractVector` with `AbstractStateSpaceSet`

This is arguably the most important change I do in this PR. Read the docstring of `StateSpaceSet` to understand why.

* Fix the `extremal_index_sueveges` function...

There were a lot of problems:

- the function name was different from where it was called
- the function input arguments were named differently from where the function was called. This was confusing.
- The order of arguments and the documentation string where wrong. This was ULTRA confusing. The threshold and quantile arguments were exchanging places in the documentation.
- The quantile and threshold were named the same way. Not possible to deduce which was what besides going through the algorithm.

* rename variables for clarity

* don't call extremal index; function is wrong

* use a normal way to get length of set

* add convenience function for dim

* first pass at performance: complete refactor of functions

The point is: local estimation should be its own function because it is a dedicated piece of algorithm. Then, we can optmize stuff by looping over each "column" of the distance matrix. If it fits in memory, we also gain massive BLAS-enabled performance. If not, we can still optimize by pre-allocating vectors and not re-allocated memory in each pass.

* add convenience function for local values

* add progress bar

and also stop using `AbstractStateSpaceSet` in function call signatures

* add test file

* correctly use greater-equals instead of only greater

* add second estimator for pareto sigma

* parallelize via threads

* add preliminary tests for extreme value dim

* fix test for extremal index

* wip] add type for deciding estimator (?)

* we don't need a type in the end.

* rename q to p as it quantifies a probability, not the quantile

* wip evt

* finish docs EVT

* add passing tests to EVT

* add to docs

* changelog

* added the functions that estimate parameters for a GEV fit.

* Added method to obtain the local dimensions and the persistances through
the Block Maxima approach

* Changes requested, removed spaces from docstring and modified the
function to get the second moment to use less memory

* Removed spaces from docstring

* Removed space between docstring and function

* The docstring is expandend and fixed according to the comments. Also the
dimension name is changed from D1 to Delta

* An empty line is removed between the docstring and the function

* Comments resolved: remove unhelpfull message, remove indexes from the
argument of the lenght function and remove the vector wrapping in the
return statement

* Changed the variable name quanti--> p like in the rest of the code

* Changed the floor function to give the correct type instead of calling
Int64 to turn argument into the desired type

* Changed the floor function to return the correct type

* Added absolute value function to prevent the domain errors in the square
root

* Added progress bar

* Added the input parameter blocksize to allow custom block size and
number of blocks. Added error handling for the case in which there is
duplicate points

* Changed the docstring to include recent changes

* Changed the error handling, created the struct BlockMaximaMM necessary for the
multiple dispatch and changed the function name

* Added the changes to include structs BlockMaxima, and Exceedances for
the multiple dispatch.
Erased the matrix allocation functions and calls.

* In the test file: Modified the tests to include the types of the new code. Include the

estimators as instances of structs and erased the test for wrong
estimator since is no longer relevant
In the extreme dims file: Update the functions so they no longer trigger errors on execution, erased the calls that still depended on estimator as a symbol instead of a struct.

* Added docstrings for the structs and fixed a couple of errors regarding
the use of structs in the code, and the call to the function if the
argument is not a struct but a Real

* Erased some comments to myself from the code

* Updated the docstrings of several functions to reflect the changes to the struct based definition of the methods

* Added back the test of the method :pwm as it was before

* Fixed running errors

Fixed errors on code: typo `x`->`X` in both functions of the gev file and in the file extremesdim changed the definition of logdist in the first function because it rewrited the vector in a way that lost elements iteratively in combination whith erasing the Inf value

* Adjusted the criteria of accuracy of the "Validity" test because the
test sometimes fails spuriously.

* Fixed the docstring of the function extremevaltheory_dim() to include
doc references and correct the missuse of the word type.

* Fixed the docstring of the function extremevaltheory_dims_persistances() to include the doc references and correct the misuse of the word type

* Corrected the misuse of type as a type.

* Erased the duplicated function

* Added reference for the block maxima extreme value method.

* add the types to the documentation

* separate docs into more blocks

* simplify docstrings of convenience functions

* separate types into their files

* create more logical structuring in the files

* fully separate code: dispatch happens in estimator files

* improve, simplify, and cohere, the codstrings

* changelog

* add more code authors

---------

Co-authored-by: Datseris <datseris.george@gmail.com>